+++ tests/test_utils.py
@@ -1,6 +1,7 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 # Test for utils
+import os
 from inspect import signature
 
 import numpy as np
@@ -8,6 +9,7 @@ import xarray as xr
 
 from xclim.core.indicator import Daily
 from xclim.core.utils import ensure_chunk_size, walk_map, wrapped_partial
+from xclim.testing import open_dataset
 
 
 def test_walk_map():
@@ -87,3 +89,10 @@ def test_ensure_chunk_size():
     assert out.chunks[0] == (3, 3, 3, 3, 3, 5)
     assert out.chunks[1] == (10, 11)
     assert out.chunks[2] == (20,)
+
+
+def test_open_testdata():
+    ds = open_dataset(
+        os.path.join("cmip5", "tas_Amon_CanESM2_rcp85_r1i1p1_200701-200712")
+    )
+    assert ds.lon.size == 128